Module mmtk::util::conversions

Calculation, conversion and rounding for memory related numbers.

Functions

address_to_chunk_index

bytes_to_chunks_up

bytes_to_formatted_string

Convert size in bytes to a readable short string, such as 1GB, 2TB, etc. It only keeps the major unit and keeps no fraction.

bytes_to_pages

bytes_to_pages_up

Convert size in bytes to the number of pages (aligned up)

chunk_align_down

chunk_align_up

chunk_index_to_address

is_address_aligned

is_page_aligned

mmap_chunk_align_down

mmap_chunk_align_up

page_align_down

pages_to_bytes

raw_align_down

raw_align_up

raw_is_aligned